# [Sofa.Type] Fix/clean and speed up of Mat