--- a/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/ProjectToLineConstraint.inl
+++ b/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/ProjectToLineConstraint.inl
@@ -161,7 +161,7 @@ void  ProjectToLineConstraint<DataTypes>::updateJacobian()
         }
         else           // unconstrained particle: set diagonal to identity block
         {
-            jacobian.insertBackBlock(i,i,Block::s_identity);
+            jacobian.insertBackBlock(i,i,Block::Identity());
         }
         i++;
     }