--- a/Sofa/framework/Type/src/sofa/type/Mat.h
+++ b/Sofa/framework/Type/src/sofa/type/Mat.h
@@ -1251,4 +1251,25 @@ constexpr Mat<3,3,real> multTranspose(const Mat<3,3,real>& m1, const Mat<3,3,rea
     return r;
 }
 
+#if !defined(SOFA_TYPE_MAT_CPP)
+
+extern template class SOFA_TYPE_API Mat<2, 2, SReal>;
+extern template class SOFA_TYPE_API Mat<2, 3, SReal>;
+extern template class SOFA_TYPE_API Mat<3, 3, SReal>;
+extern template class SOFA_TYPE_API Mat<4, 4, SReal>;
+extern template class SOFA_TYPE_API Mat<6, 3, SReal>;
+extern template class SOFA_TYPE_API Mat<6, 6, SReal>;
+extern template class SOFA_TYPE_API Mat<8, 3, SReal>;
+extern template class SOFA_TYPE_API Mat<8, 8, SReal>;
+extern template class SOFA_TYPE_API Mat<9, 9, SReal>;
+extern template class SOFA_TYPE_API Mat<12, 3, SReal>;
+extern template class SOFA_TYPE_API Mat<12, 6, SReal>;
+extern template class SOFA_TYPE_API Mat<12, 12, SReal>;
+extern template class SOFA_TYPE_API Mat<20, 20, SReal>;
+extern template class SOFA_TYPE_API Mat<20, 32, SReal>;
+extern template class SOFA_TYPE_API Mat<24, 24, SReal>;
+extern template class SOFA_TYPE_API Mat<32, 20, SReal>;
+
+#endif
+
 } // namespace sofa::type