@alxbilger 2022-09-07
Before the changes:

```
------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations
------------------------------------------------------------------------------
BM_Matrix_typematf_matmult<3>/128         1.27 us         1.28 us       560000
BM_Matrix_typematf_matmult<3>/256         2.53 us         2.51 us       280000
BM_Matrix_typematf_matmult<3>/512         5.11 us         5.16 us       100000
BM_Matrix_eigenmatf_matmult<3>/128       0.923 us        0.900 us       746667
BM_Matrix_eigenmatf_matmult<3>/256        1.83 us         1.84 us       373333
BM_Matrix_eigenmatf_matmult<3>/512        3.67 us         3.68 us       186667
```

After the changes:

```
------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations
------------------------------------------------------------------------------
BM_Matrix_typematf_matmult<3>/128         1.00 us         1.00 us       640000
BM_Matrix_typematf_matmult<3>/256         1.98 us         1.97 us       373333
BM_Matrix_typematf_matmult<3>/512         3.93 us         3.90 us       172308
BM_Matrix_eigenmatf_matmult<3>/128        1.00 us         1.00 us       746667
BM_Matrix_eigenmatf_matmult<3>/256        1.99 us         2.01 us       373333
BM_Matrix_eigenmatf_matmult<3>/512        3.99 us         4.01 us       179200
```