@alxbilger 2022-09-08
> It does not come from your PR, but while trying to compile this branch with the newest gcc, there is an interesting (and useful) warning:
> 
> ```
> /home/fred/sofa/src/current/Sofa/framework/Type/src/sofa/type/Mat.h:408:58: warning: array subscript 2 is above array bounds of ‘const sofa::type::VecNoInit<3, double> [2]’ [-Warray-bounds]
>   408 |                 if( rabs( this->elems[i][j] - this->elems[j][i] ) > EQUALITY_THRESHOLD ) return false;
> ```
> 
> Indeed, there is no compile-time test about the fact that the matrix is square or not (so the symmetry could be irrelevant)

I added compile-time check that the matrix is square. Thanks for spotting it.