@alxbilger 2022-09-08
Now the  `multTranspose` operation is faster than Eigen:

```
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
BM_Matrix_typemat3x3f_multTranspose/128       0.731 us        0.711 us       746667
BM_Matrix_typemat3x3f_multTranspose/256        1.53 us         1.54 us       497778
BM_Matrix_typemat3x3f_multTranspose/512        2.83 us         2.85 us       235789
BM_Matrix_eigenmat3x3f_multTranspose/128      0.842 us        0.837 us       746667
BM_Matrix_eigenmat3x3f_multTranspose/256       1.73 us         1.73 us       407273
BM_Matrix_eigenmat3x3f_multTranspose/512       3.46 us         3.45 us       194783

```