@alxbilger 2022-09-08
New benchmark:

```
------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations
------------------------------------------------------------------------------
BM_Matrix_typematf_matmult<3>/128        0.689 us        0.684 us      1120000
BM_Matrix_typematf_matmult<3>/256         1.28 us         1.29 us       497778
BM_Matrix_typematf_matmult<3>/512         2.76 us         2.73 us       263529
BM_Matrix_eigenmatf_matmult<3>/128        1.01 us         1.03 us       746667
BM_Matrix_eigenmatf_matmult<3>/256        1.93 us         1.95 us       344615
BM_Matrix_eigenmatf_matmult<3>/512        3.81 us         3.81 us       172308
```

We are faster than Eigen!