@damienmarchal 2022-09-21
Ready, merge if [ci-build][with-all-tests] agrees.