@alxbilger 2022-09-21
Two unit tests failing, probably solved in https://github.com/sofa-framework/sofa/pull/3322